+++ src/main/java/ro/danix/first/model/repository/user/mongo/UserRepositoryImpl.java
@@ -1,9 +1,9 @@
 package ro.danix.first.model.repository.user.mongo;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import static org.springframework.data.mongodb.core.query.Criteria.*;
 import static org.springframework.data.mongodb.core.query.Query.*;
-
+import java.math.BigDecimal;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.mongodb.core.MongoOperations;
 import org.springframework.data.mongodb.core.query.Query;
 import org.springframework.stereotype.Repository;
@@ -28,7 +28,7 @@ public class UserRepositoryImpl implements UserRepository {
     }
 
     @Override
-    public User findOne(Long id) {
+    public User findOne(BigDecimal id) {
         Query query = query(where("id").is(id));
         return operations.findOne(query, User.class);
     }
